[web] Rename file name: `@/dashboard/types/balance` -> `@/dashboard/types/WalletBalance`
